backend/pyproject.toml
[tool.poetry]
name = "Digital Lions"
version = "0.1.0"
description = "Backend API for Digital Lions"
authors = ["Anne Lohmeijer <a.lohmeijer@hotmail.com>"]
readme = "README.md"
packages = [
    { include = "app"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
psycopg2-binary = "2.9.5"
sqlmodel = "^0.0.18"
pydantic = {extras = ["email"], version = "^2.7.1"}
bcrypt = "^4.1.3"
alembic = "^1.13.1"
pydantic-settings = "^2.3.0"
jinja2 = "^3.1.4"
pyjwt = "^2.9.0"
fastapi-mail = "^1.4.1"
resend = "^2.4.0"
cryptography = "^43.0.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.3.4"
pytest-cov = "^5.0.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.11.1"
pytest-env= "^1.1.3"

faker = "^30.2.0"
[tool.poetry.scripts]
start = "python run app.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["app"]

[tool.pytest_env]
ALLOWED_ORIGINS="http://localhost:8000,http://digitallions.com,https://digital-lion.com"
FEATURE_API_KEY="false"
POSTGRES_DATABASE_URL="postgresql://postgres:postgres@localhost:5432/digital_lion_test"

[tool.ruff]
extend-exclude = ["alembic/*"]
fix = true
show-fixes = true
line-length = 99
target-version = "py311"

[tool.black]
extend-exclude = '''alembic/*'''

[tool.ruff.lint]
ignore = []
select = ["E", "F", "W", "I", "UP", "Q", "TID"]
